Restructure sources, rename plot API, and clean macOS artifacts

+++ solve_sq_roots/headers/solve.h
@@ -5,7 +5,7 @@ int solve_square_equation(double a, double b, double c, double *x_1, double *x_2
 
 int solve_full_square_equation(double a, double b, double c, double *x_1, double *x_2);
 
-int find_discriminant(double a, double b, double c);
+double find_discriminant(double a, double b, double c);
 
 int solve_linear_equation(double k, double b, double *x_1);
